cases: eth: Make sure both ifaces link is up
It isn't guaranteed that the first interface link is getting up
synchronously at the time the second interface link is setup. In
this case we must make sure that both of them have the operstate
'up' value after the links setting up procedure.

It is also a good practice to perform a cleanup procedure in a
reverse order. This is done for the links enable/disable commands.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>