cases: serial_loopback: Add full COM-port config
Some ports may expose a full modem-control pinout. In this case they can
be tested using the full set of the pins layout. This can be done by
enabling '-clocal crtscts' stty configs. Lets add a 'modem' boolean
parameter of the serial_loopback test, which would perform the stty-based
setup of the ports marked as supporting the full COM-port layout.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>